# rock-platform/rock/docs/swagger.yaml
definitions:
  api.UserDetailResp:
    properties:
      created_at:
        example: "2020-12-20 15:15:22"
        type: string
      email:
        example: admin_user@sensetime.com
        type: string
      id:
        description: type UserFullResp struct {
        example: 1
        type: integer
      login_block_until:
        example: "2020-12-20 15:15:22"
        type: string
      name:
        example: admin_user
        type: string
      password:
        example: '********'
        type: string
      role_created_at:
        example: "2020-12-20 15:15:22"
        type: string
      role_description:
        example: description for role
        type: string
      role_id:
        type: integer
      role_name:
        type: string
      role_updated_at:
        example: "2020-12-20 15:15:22"
        type: string
      role_version:
        example: 1
        type: integer
      salt:
        example: salt secret
        type: string
      token:
        example: user token
        type: string
      updated_at:
        example: "2020-12-20 15:15:22"
        type: string
    required:
    - role_description
    - role_version
    type: object
...
  utils.HTTPError:
    properties:
      error:
        example: error message here
        type: string
      error_code:
        example: 50000001
        type: integer
    type: object
...
  v1.CreateRoleReq:
    properties:
      description:
        example: description for role
        type: string
      name:
        example: admin_role
        type: string
    required:
    - name
    type: object
...
  v1.LoginUserInfo:
    properties:
      name:
        example: admin_user
        type: string
      password:
        example: '********'
        type: string
    required:
    - name
    - password
    type: object
  v1.PaginateBriefUserResp:
    properties:
      items:
        items:
          $ref: '#/definitions/v1.UserBriefResp'
        type: array
      page_num:
        example: 1
        type: integer
      pages:
        example: 1
        type: integer
      per_size:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    required:
    - items
    - page_num
    - pages
    - per_size
    - total
    type: object
  v1.PaginateRoleResp:
    properties:
      items:
        items:
          $ref: '#/definitions/v1.RoleBriefResp'
        type: array
      page_num:
        example: 1
        type: integer
      pages:
        example: 1
        type: integer
      per_size:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    required:
    - items
    - page_num
    - pages
    - per_size
    - total
    type: object
...
  v1.ResetPwdReq:
    properties:
      email:
        example: admin_user@sensetime.com
        type: string
      password:
        example: '********'
        type: string
      re_password:
        example: '********'
        type: string
      secret:
        example: TR6UdhT7ebJOCC5N
        type: string
    required:
    - email
    - password
    - re_password
    - secret
    type: object
  v1.ResetUserReq:
    properties:
      email:
        example: admin_user@sensetime.com
        type: string
    required:
    - email
    type: object
  v1.RoleBriefResp:
    properties:
      created_at:
        example: "2018-10-09T14:57:23+08:00"
        type: string
      description:
        example: description for role
        type: string
      id:
        example: 1
        type: integer
      name:
        example: admin_role
        type: string
      updated_at:
        example: "2018-10-09T14:57:23+08:00"
        type: string
      version:
        example: 1
        type: integer
    required:
    - created_at
    - description
    - id
    - name
    - updated_at
    - version
    type: object
  v1.UpdateRoleReq:
    properties:
      description:
        example: description for role
        type: string
    type: object
...
  v1.UserBriefResp:
    properties:
      created_at:
        example: "2018-10-09T14:57:23+08:00"
        type: string
      email:
        example: admin@sensetime.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: admin_role
        type: string
      role_id:
        example: 1
        type: integer
      updated_at:
        example: "2018-10-09T14:57:23+08:00"
        type: string
      version:
        example: 1
        type: integer
    required:
    - created_at
    - email
    - id
    - name
    - role_id
    - updated_at
    - version
    type: object
info:
  contact: {}
paths:
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Api to login rock platform with name and password
      parameters:
      - description: JSON type input body
        in: body
        name: input_body
        required: true
        schema:
          $ref: '#/definitions/v1.LoginUserInfo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.UserDetailResp'
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Login rock platform with name and password
      tags:
      - AUTH
  /v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Api to logout rock platform
      produces:
      - application/json
      responses:
        "204":
          description: StatusNoContent
          schema:
            type: string
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Logout rock platform
      tags:
      - AUTH
  /v1/auth/pwd:
    post:
      consumes:
      - application/json
      description: Api to update user password with secret
      parameters:
      - description: JSON type input body
        in: body
        name: input_body
        required: true
        schema:
          $ref: '#/definitions/v1.ResetPwdReq'
      produces:
      - application/json
      responses:
        "200":
          description: StatusOK
          schema:
            type: string
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Update user's password with secret
      tags:
      - AUTH
  /v1/auth/reset:
    post:
      consumes:
      - application/json
      description: Api to create reset email
      parameters:
      - description: JSON type input body
        in: body
        name: input_body
        required: true
        schema:
          $ref: '#/definitions/v1.ResetUserReq'
      produces:
      - application/json
      responses:
        "204":
          description: StatusNoContent
          schema:
            type: string
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Create reset email
      tags:
      - AUTH
...
  /v1/roles:
    get:
      consumes:
      - application/json
      description: api for get all roles
      parameters:
      - default: 1
        description: Request page number
        in: query
        name: page_num
        required: true
        type: integer
      - default: 10
        description: Request page size
        in: query
        name: page_size
        required: true
        type: integer
      - description: 'Fuzzy Query(field: name)'
        in: query
        name: QueryField
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: StatusOK
          schema:
            $ref: '#/definitions/v1.PaginateRoleResp'
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Get role
      tags:
      - ROLE
    post:
      consumes:
      - application/json
      description: api for create role
      parameters:
      - description: JSON type input body
        in: body
        name: input_body
        required: true
        schema:
          $ref: '#/definitions/v1.CreateRoleReq'
      produces:
      - application/json
      responses:
        "201":
          description: StatusCreated
          schema:
            $ref: '#/definitions/v1.RoleBriefResp'
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Create role
      tags:
      - ROLE
  /v1/roles/{id}:
    delete:
      consumes:
      - application/json
      description: api for get an role by id
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: StatusNoContent
          schema:
            type: string
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Get an role by id
      tags:
      - ROLE
    get:
      consumes:
      - application/json
      description: api for get an role by id
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: StatusOK
          schema:
            $ref: '#/definitions/v1.RoleBriefResp'
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Get an role by id
      tags:
      - ROLE
    put:
      consumes:
      - application/json
      description: api for update role description
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role Description
        in: body
        name: description
        required: true
        schema:
          type: string
      - description: JSON type for update role description
        in: body
        name: update_body
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateRoleReq'
      produces:
      - application/json
      responses:
        "200":
          description: StatusOK
          schema:
            $ref: '#/definitions/v1.RoleBriefResp'
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Update role description by id and body
      tags:
      - ROLE
  /v1/roles/{id}/users:
    get:
      consumes:
      - application/json
      description: api for get all users by species role id
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Request page number
        in: query
        name: page_num
        required: true
        type: integer
      - default: 10
        description: Request page size
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: StatusOK
          schema:
            $ref: '#/definitions/v1.PaginateBriefUserResp'
        "400":
          description: StatusBadRequest
          schema:
            $ref: '#/definitions/utils.HTTPError'
        "500":
          description: StatusInternalServerError
          schema:
            $ref: '#/definitions/utils.HTTPError'
      summary: Get all users by species role id
      tags:
      - ROLE
...
swagger: "2.0"